server/pyproject.toml
[tool.poetry]
name = "photodetect-server"
version = "0.1.0"
description = ""
authors = ["AlekseyAmp <thientaovv@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.110.0"
uvicorn = "^0.28.0"
python-dotenv = "^1.0.1"
pydantic = "^2.6.4"
alembic = "^1.13.1"
sqlalchemy = "^2.0.28"
isort = "^5.13.2"
flake8 = "^7.0.0"
pytz = "^2024.1"
psycopg2-binary = "^2.9.9"
ultralytics = "^8.1.29"
python-multipart = "^0.0.9"
pillow = "^10.2.0"
autopep8 = "^2.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.isort]
profile = "black"
known_first_party = "src"
known_fastapi = "fastapi"
sections=["FUTURE", "STDLIB", "THIRDPARTY", "FASTAPI", "FIRSTPARTY", "LOCALFOLDER"]
line_length = 80

[tool.flake8]
max-line-length = 80
select = ['F', 'E', 'W', ]

ignore = ['W503', ]

per-file-ignores = [
    '*/adapters/*/repositories.py:E711,E712',
    '*/adapters/*/repositories/*:E711,E712',
    '*/migrations/*:E711,E712',
    '*/alembic/versions/*:E711,E712',
    'init.py:F401',
]
count = true
